+++ crates/web-client/src/new_transactions.rs
@@ -12,8 +12,9 @@ use wasm_bindgen::prelude::*;
 
 use crate::{
     models::{
-        account_id::AccountId, note_type::NoteType, transaction_request::TransactionRequest,
-        transaction_result::TransactionResult, transactions::NewSwapTransactionResult,
+        account_id::AccountId, note_type::NoteType, provers::TransactionProver,
+        transaction_request::TransactionRequest, transaction_result::TransactionResult,
+        transactions::NewSwapTransactionResult,
     },
     WebClient,
 };
@@ -64,6 +65,26 @@ impl WebClient {
                     })?;
                 },
             }
+
+            Ok(())
+        } else {
+            Err(JsValue::from_str("Client not initialized"))
+        }
+    }
+
+    pub async fn submit_transaction_with_prover(
+        &mut self,
+        transaction_result: &TransactionResult,
+        prover: TransactionProver,
+    ) -> Result<(), JsValue> {
+        if let Some(client) = self.get_mut_inner() {
+            let native_transaction_result: NativeTransactionResult = transaction_result.into();
+            client
+                .submit_transaction_with_prover(native_transaction_result, prover.get_prover())
+                .await
+                .map_err(|err| {
+                    JsValue::from_str(&format!("Failed to submit Transaction: {}", err))
+                })?;
 
             Ok(())
         } else {