+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * [BREAKING] Added support for specifying map storage slots for FPI (#645)
 * Limited the number of decimals that an asset can have (#666).
 * [BREAKING] Removed the `testing` feature from the CLI (#670).
+* Added per transaction prover support to the web client (#674).
 
 ### Fixes